# Improve performance of numeric conversions