--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,4 +1,8 @@
-1.0.8
+1.0.10
+======
+ - Improve performance of `toInt` and `fromInt`, `toLong` and `fromLong`, etc.
+
+1.0.9
 =====
  - *Significant* performance improvement in `toBase64`.
  - Fixed source links in ScalaDoc.