## toInt

`run -i 10 -wi 10 -f1 -t1 .*NumericConv.*`

### Before

```
[info] Benchmark                                                          Mode  Cnt    Score    Error  Units
[info] NumericConversionBenchmarks.fromInt_16bit_bigEndian                avgt   10  146.476 ±  7.589  us/op
[info] NumericConversionBenchmarks.fromInt_30bit_bigEndian                avgt   10  135.761 ± 10.377  us/op
[info] NumericConversionBenchmarks.fromInt_30bit_littleEndian             avgt   10  285.830 ± 20.083  us/op
[info] NumericConversionBenchmarks.fromInt_32bit_bigEndian                avgt   10   43.472 ±  3.357  us/op
[info] NumericConversionBenchmarks.fromInt_32bit_littleEndian             avgt   10   48.617 ±  2.885  us/op
[info] NumericConversionBenchmarks.toInt_14bit_bigEndian                  avgt   10   44.556 ±  3.599  us/op
[info] NumericConversionBenchmarks.toInt_16bit_bigEndian                  avgt   10   43.517 ±  0.852  us/op
[info] NumericConversionBenchmarks.toInt_16bit_bigEndian_unsigned         avgt   10   42.875 ±  1.438  us/op
[info] NumericConversionBenchmarks.toInt_16bit_littleEndian               avgt   10   82.882 ±  4.141  us/op
[info] NumericConversionBenchmarks.toInt_16bit_littleEndian_unsigned      avgt   10   80.745 ±  1.668  us/op
[info] NumericConversionBenchmarks.toInt_24bit_bigEndian                  avgt   10   55.102 ±  1.550  us/op
[info] NumericConversionBenchmarks.toInt_32bit_bigEndian                  avgt   10   66.239 ±  2.126  us/op
[info] NumericConversionBenchmarks.toInt_32bit_bigEndian_nonCompacted     avgt   10   65.913 ±  1.748  us/op
[info] NumericConversionBenchmarks.toInt_32bit_littleEndian               avgt   10  110.445 ±  3.446  us/op
[info] NumericConversionBenchmarks.toInt_32bit_littleEndian_nonCompacted  avgt   10  133.427 ±  5.729  us/op
[info] NumericConversionBenchmarks.toInt_8bit_bigEndian                   avgt   10   43.952 ±  1.666  us/op
[info] NumericConversionBenchmarks.toLong_32bit_bigEndian                 avgt   10   77.413 ±  1.844  us/op
[info] NumericConversionBenchmarks.toLong_60bit_bigEndian                 avgt   10  113.957 ±  4.130  us/op
[info] NumericConversionBenchmarks.toLong_64bit_bigEndian                 avgt   10  116.911 ± 11.796  us/op
[info] NumericConversionBenchmarks.toLong_64bit_littleEndian              avgt   10  199.810 ±  4.788  us/op
```

### After

```
[info] Benchmark                                                          Mode  Cnt    Score    Error  Units
[info] NumericConversionBenchmarks.fromInt_16bit_bigEndian                avgt   10  124.248 ± 11.345  us/op
[info] NumericConversionBenchmarks.fromInt_30bit_bigEndian                avgt   10  147.310 ± 10.509  us/op
[info] NumericConversionBenchmarks.fromInt_30bit_littleEndian             avgt   10  288.634 ±  5.093  us/op
[info] NumericConversionBenchmarks.fromInt_32bit_bigEndian                avgt   10   28.400 ±  0.798  us/op
[info] NumericConversionBenchmarks.fromInt_32bit_littleEndian             avgt   10   48.125 ±  1.913  us/op
[info] NumericConversionBenchmarks.toInt_14bit_bigEndian                  avgt   10   58.432 ±  2.919  us/op
[info] NumericConversionBenchmarks.toInt_16bit_bigEndian                  avgt   10   48.174 ±  1.982  us/op
[info] NumericConversionBenchmarks.toInt_16bit_bigEndian_unsigned         avgt   10   46.671 ±  1.918  us/op
[info] NumericConversionBenchmarks.toInt_16bit_littleEndian               avgt   10   48.120 ±  2.634  us/op
[info] NumericConversionBenchmarks.toInt_16bit_littleEndian_unsigned      avgt   10   47.902 ±  3.283  us/op
[info] NumericConversionBenchmarks.toInt_24bit_bigEndian                  avgt   10   68.415 ±  2.728  us/op
[info] NumericConversionBenchmarks.toInt_32bit_bigEndian                  avgt   10   47.594 ±  1.152  us/op
[info] NumericConversionBenchmarks.toInt_32bit_bigEndian_nonCompacted     avgt   10   49.274 ±  2.916  us/op
[info] NumericConversionBenchmarks.toInt_32bit_littleEndian               avgt   10   48.272 ±  1.832  us/op
[info] NumericConversionBenchmarks.toInt_32bit_littleEndian_nonCompacted  avgt   10   48.003 ±  1.190  us/op
[info] NumericConversionBenchmarks.toInt_8bit_bigEndian                   avgt   10   43.301 ±  1.309  us/op
[info] NumericConversionBenchmarks.toLong_32bit_bigEndian                 avgt   10   34.476 ±  4.187  us/op
[info] NumericConversionBenchmarks.toLong_60bit_bigEndian                 avgt   10   89.732 ±  5.049  us/op
[info] NumericConversionBenchmarks.toLong_64bit_bigEndian                 avgt   10   31.587 ±  1.851  us/op
[info] NumericConversionBenchmarks.toLong_64bit_littleEndian              avgt   10   31.675 ±  0.892  us/op
```